refactor: introduced `Builder` and `JarBuilder`
