+++ src/main/java/dev/jbang/cli/Export.java
@@ -1,9 +1,9 @@
 package dev.jbang.cli;
 
 import static dev.jbang.cli.BaseBuildCommand.buildIfNeeded;
-import static dev.jbang.cli.BaseBuildCommand.getImageName;
 import static dev.jbang.cli.BaseScriptCommand.enableInsecure;
 import static dev.jbang.cli.Export.handle;
+import static dev.jbang.source.JarBuilder.getImageName;
 import static dev.jbang.util.JavaUtil.resolveInJavaHome;
 import static dev.jbang.util.Util.downloadFile;
