+++ src/main/java/dev/jbang/source/Builder.java
@@ -1,0 +1,9 @@
+package dev.jbang.source;
+
+import java.io.IOException;
+
+public interface Builder {
+	Source build(ScriptSource src, RunContext ctx) throws IOException;
+
+	Builder nop = (src, ctx) -> null;
+}
